Authentication for all routes

+++ backend/src/controllers/goalController.ts
@@ -1,12 +1,13 @@
 import { Request, Response } from 'express';
 import asyncHandler from 'express-async-handler';
 import { Goal } from '../models/goalModel';
+import { User } from '../models/userModel';
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 export const getGoals = asyncHandler(async (req: Request, res: Response) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.body.user.id });
 
   res.status(200).json(goals);
 });
@@ -22,6 +23,7 @@ export const setGoal = asyncHandler(async (req: Request, res: Response) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.body.user.id,
   });
 
   res.status(200).json(goal);
@@ -36,6 +38,20 @@ export const updateGoal = asyncHandler(async (req: Request, res: Response) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.body.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Verify logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error('User not authorized');
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -54,6 +70,20 @@ export const deleteGoal = asyncHandler(async (req: Request, res: Response) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.body.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Verify logged in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(400);
+    throw new Error('User not authorized');
   }
 
   await goal.deleteOne();
